split done

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: ael-mank <ael-mank@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 08:49:15 by ael-mank          #+#    #+#             */
-/*   Updated: 2023/11/10 23:36:42 by ael-mank         ###   ########.fr       */
+/*   Updated: 2023/11/11 20:04:14 by ael-mank         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
